Extract test utils
