+++ tlspuffin/src/tls/vulnerabilities.rs
@@ -1097,6 +1097,7 @@ pub mod tests {
 
     use crate::{
         put_registry::{TLS_PUT_REGISTRY, WOLFSSL520_PUT},
+        test_utils::expect_crash,
         tls::{
             seeds::{
                 seed_client_attacker12, seed_client_attacker_auth, seed_client_attacker_full,
@@ -1142,34 +1143,6 @@ pub mod tests {
                     Action::Output(_) => {}
                 }
             }
-        }
-    }
-
-    fn expect_crash<R>(mut func: R)
-    where
-        R: FnMut(),
-    {
-        match unsafe { fork() } {
-            Ok(ForkResult::Parent { child, .. }) => {
-                let status = waitpid(child, Option::from(WaitPidFlag::empty())).unwrap();
-
-                if let Signaled(_, signal, _) = status {
-                    if signal != Signal::SIGSEGV && signal != Signal::SIGABRT {
-                        panic!("Trace did not crash with SIGSEGV/SIGABRT!")
-                    }
-                } else if let Exited(_, code) = status {
-                    if code == 0 {
-                        panic!("Trace did not crash exit with non-zero code (AddressSanitizer)!")
-                    }
-                } else {
-                    panic!("Trace did not signal!")
-                }
-            }
-            Ok(ForkResult::Child) => {
-                func();
-                std::process::exit(0);
-            }
-            Err(_) => panic!("Fork failed"),
         }
     }
